requirements.txt
boto3
python-dotenv
flask-cors